test /3

+++ .ci/travis-checks.sh
@@ -1,28 +1,32 @@
 #!/bin/bash
 
 if [[ $TRAVIS_OS_NAME == "linux" ]]; then
-  echo "Checking code style..."
-  pwd
-  clang-format -i \
-    common/*.h \
-    common/*.cpp \
-    cockatrice/src/*.h \
-    cockatrice/src/*.cpp \
-    oracle/src/*.h \
-    oracle/src/*.cpp \
-    servatrice/src/*.h \
-    servatrice/src/*.cpp
-    
-  git clean -f
-  git diff --quiet || (
-    echo "*****************************************************";
-    echo "***  This PR is not clean against our code style  ***";
-    echo "***  Run clang-format and fix up any differences  ***";
-    echo "***  Check our CONTRIBUTING.md file for details!  ***";
-    echo "***                  Thank you ♥                  ***";
-    echo "*****************************************************";
-  )
-  git diff --exit-code
+  if [[ $BUILDTYPE == "Debug" ]]; then #TODO: generalize h cpp files search to top-directory? / check with ctrlaltca
+    echo "Checking code style..."
+    clang-format -i \
+      common/*.h \
+      common/*.cpp \
+      cockatrice/src/*.h \
+      cockatrice/src/*.cpp \
+      oracle/src/*.h \
+      oracle/src/*.cpp \
+      servatrice/src/*.h \
+      servatrice/src/*.cpp
+      
+    git clean -f
+    git diff --quiet && echo "Code style - OK" || (
+      echo "*****************************************************";
+      echo "***  This PR is not clean against our code style  ***";
+      echo "***  Run clang-format and fix up any differences  ***";
+      echo "***  Check our CONTRIBUTING.md file for details!  ***";
+      echo "***                  Thank you ♥                  ***";
+      echo "*****************************************************";
+    )
+    git diff --exit-code
+  else
+    echo "Skipping... checks are only performed on runs in debug mode"
 fi
 
-#TODO: message on success
+if [[ $TRAVIS_OS_NAME == "osx" ]]; then
+  echo "Skipping... checks are only performed on runs from Linux"
+fi
